+++ libraries/default.rb
@@ -1,8 +1,8 @@
 # Chrome helper
 module Chrome
   def chrome_version
+    return unless platform?('windows')
     install_dir = node['chrome']['master_preferences'].slice(0, node['chrome']['master_preferences'].rindex('\\'))
-    install_dir = 'c:\selenium2'
     entries = ::Dir.entries(install_dir).reverse
     entries.each do |name|
       return name if name[/^[\d|.].*$/]