+++ CHANGELOG.md
@@ -1,8 +1,14 @@
-# Chrome CHANGELOG
+# Chrome Changelog
+
+## 0.3.0
+
+- Add chrome_version method
+- Install Chrome at compile time
+- Warn if not windows platform
 
 ## 0.2.0
 
-- Added preferences resource
+- Add preferences resource
 
 ## 0.1.0